capital_v

+++ routes/web.php
@@ -13,7 +13,8 @@ Route::get('/', [PestañaHomeController::class, 'vista_home_user'])->name('home'
 
 Route::get('/direccion', function () {return view('usuario.Organizacion.Direccion');})->name('direccion');
 
-Route::get('/capital', function () {return view('usuario.Organizacion.CapitalHumano');})->name('capital');
+Route::get('/capitales', [CapitalHumanoController::class, 'capHumano_us'])->name('capital_usuario');
+
 Route::get('/areas', function () {return view('usuario.Organizacion.AreasInvestigacion');})->name('areas');
 
 // RUTAS DE SECCIÓN NOSOTROS
@@ -193,12 +194,12 @@ Route::delete('/admin/noticias/{id}', [NoticiaController::class, 'destroy'])->na
 // ------------------------- CRUD ORGANIZACION ---------------------------------------------
 // ---- Capital Humano ---- //
 
-Route::get('/capitales', [CapitalHumanoController::class, 'index'])->name('capital_index');
 Route::get('/admin/capital_humano', [CapitalHumanoController::class, 'index'])->name('capital_index');
-Route::post('/capitales', [CapitalHumanoController::class, 'store'])->name('capitales.store');
-Route::get('capitales/{id}/edit', [CapitalHumanoController::class, 'edit'])->name('capitales.edit');
-Route::put('capitales/{id}', [CapitalHumanoController::class, 'update'])->name('capitales.update');
-Route::delete('capitales/{id}', [CapitalHumanoController::class, 'destroy'])->name('capitales.destroy');
+
+Route::post('/admin/capitales', [CapitalHumanoController::class, 'store'])->name('capitales.store');
+Route::get('/admin/capitales/{id}/edit', [CapitalHumanoController::class, 'edit'])->name('capitales.edit');
+Route::put('/admin/capitales/{id}', [CapitalHumanoController::class, 'update'])->name('capitales.update');
+Route::delete('/admin/capitales/{id}', [CapitalHumanoController::class, 'destroy'])->name('capitales.destroy');
